fix(audit): 48 minor-severity bugs + dead-code removal
Third batch from the whole-codebase audit (per-file apply, then adversarial
verify over the diffs -- 27/27 clean; util.c hand-verified after an agent
misfired its edit into the stale worktree, which was caught and ported here;
an out-of-scope IOP ziso_align change was reverted). Highlights:

- system.c: USBA_crc32 indexed its table with a signed char (negative -> OOB read).
- config.c: composedKey buffer was char[32] but holds prefix_key (up to 63) ->
  truncated/unreadable round-trip; widened to [64]. Also guarded a malformed
  IPCONFIG sscanf and dropped an implementation-defined pointer->int cast.
- dia.c: cancel-while-editing a UI_ENUM didn't restore the prior value; and the
  spurious-(-1) guard fired for any dialog whose first item is a UI_HEADER (id 0).
- fntsys.c: kerning was permanently dead (previous glyph never recorded) -> enabled.
- bdmsupport.c: a missing %d (silent), and an all-slots-present case that fell into
  a full unnecessary DelayThread stall before the rescan.
- pad.c: analog-stick-only nav didn't set rcode/reset the screensaver; bounds.
- opl.c: 3 auto-launch malloc sites now NULL-checked.
- util.c: readFileBuffer read()<0 -> OOB write guard; sysDeleteFolder now propagates
  recursive-delete failures; removed dead logfile()/logbuffer().
- ethsupport.c: empty SMB share-list left stale phantom entries on screen.
- texcache.c: deleted 3 unused static helpers (unused-function warnings, ELF bloat).
- vmc_groups.c: deleted the dead duplicate XEBP_100.72 group; opl/dialogs/menusys/
  mmcesupport/hdd/httpclient/sound/zso/cheatman/guigame/OSDHistory/supportbase/debug
  each got a real leak/UAF/bounds/return-value/uninitialized/dead-code fix.
- include/debug.h, include/supportbase.h: signature + ghost-declaration cleanups.

Clean build green (header changes -> full rebuild).

+++ src/cheatman.c
@@ -182,9 +182,9 @@ static inline int is_cmt_str(const char *s)
  * chr_idx - Returns the index within @s of the first occurrence of the
  * specified char @c.  If no such char occurs in @s, then (-1) is returned.
  */
-static size_t chr_idx(const char *s, char c)
+static int chr_idx(const char *s, char c)
 {
-    size_t i = 0;
+    int i = 0;
 
     while (s[i] && (s[i] != c))
         i++;
